requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0  # optional
accelerate>=0.24.0
safetensors>=0.4.0
einops>=0.7.0